Give up trying to enable APIs through terraform
I'll just do it with gcloud. terraform disables the APIs on destroy,
which fails because the cloud resource API gets disabled, which then
blocks terraform from disabling the other APIs. We don't want any APIs
disabled! These APIs are too big of a pain in the ass. So I'll just
enable them at with gcloud and leave them enabled forever!